requirements.txt
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.1
markdown>=3.6
playwright>=1.45
